MMT-2067 updated tests
