+++ spec/features/manage_proposals/manage_proposal_publish_spec.rb
@@ -1,6 +1,7 @@
 describe 'When publishing collection draft proposals', js: true do
   before do
     login(real_login: true)
+    mock_urs_get_users
     allow_any_instance_of(PermissionChecking).to receive(:is_non_nasa_draft_approver?).and_return(true)
   end
