+++ spec/support/proposal_status_helper.rb
@@ -11,7 +11,7 @@ module Helpers
       record.proposal_status = 'submitted'
       record.status_history = { 'submitted' => { 'username' => 'TestUser1', 'action_date' => '2019-10-11 01:00' } }
       record.approver_feedback = {}
-      record.submitter_id = 'TestUser1'
+      record.submitter_id = 'testuser'
       record.save
     end
 
@@ -23,7 +23,7 @@ module Helpers
           'approved' =>
             { 'username' => 'TestUser2', 'action_date' => '2019-10-11 02:00' } }
       record.approver_feedback = {}
-      record.submitter_id = 'TestUser1'
+      record.submitter_id = 'testuser'
       record.save
     end
 
@@ -37,7 +37,7 @@ module Helpers
       record.approver_feedback =
         { 'reasons' => ['Misspellings/Grammar', 'Other'],
           'note' => 'Test Reason for rejecting a proposal' }
-      record.submitter_id = 'TestUser1'
+      record.submitter_id = 'testuser'
       record.save
     end
 
@@ -55,7 +55,7 @@ module Helpers
       record.proposal_status = 'done'
       record.status_history = { 'submitted' => { 'username' => 'TestUser1', 'action_date' => '2019-10-11 01:00' }, 'approved' => { 'username' => 'TestUser2', 'action_date' => '2019-10-11 02:00' }, 'done' => { 'username' => 'TestUser4', 'action_date' => '2019-10-11 04:00' } }
       record.approver_feedback = {}
-      record.submitter_id = 'TestUser1'
+      record.submitter_id = 'testuser'
       record.save
     end
   end
